pyproject.toml
[project]
name = "thona-etl"
version = "0.1.0"
description = "ETL Engine (Oracle + ThonaApi + Redshift) orchestrated by Prefect"
requires-python = ">=3.13,<3.14"
readme = "README.md"
dependencies = [
  "pydantic>=2,<3",
  "PyYAML>=6,<7",
  "python-dotenv>=1,<2",

  "httpx>=0.27,<1",

  "pandas>=2.2,<3",
  "pyarrow>=16,<20",
  "openpyxl>=3.1,<4",

  "SQLAlchemy>=2,<3",
  "psycopg[binary]>=3.1,<4",
  "oracledb>=2,<3",

  "prefect>=2,<4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "ruff>=0.5,<1",
]

[tool.ruff]
line-length = 110
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]